// app/types/Customer.tsx
import { Country } from "./Country";
import { Representative } from "./Representative";

export interface Customer {
  id: number;
  name: string;
  country: Country;
  company: string;
  date: string | Date;
  status: string;
  verified: boolean;
  activity: number;
  representative: Representative;
  balance: number;
}